changes for the last couple of weeks

+++ share/fs-uae/shaders/scanlines-classic.shader
@@ -15,13 +15,13 @@ uniform sampler2D rubyTexture;
 
 void main(void) {   
     vec4 rgb = texture2D(rubyTexture, gl_TexCoord[0].xy);
-	int line = int(gl_FragCoord.y);
-	if (mod(line, 2) == 0) {
-		gl_FragColor = mix(rgb, vec4(1.0, 1.0, 1.0, 1.0), 0.05);
-	}
-	else {
-		gl_FragColor = rgb * vec4(0.85, 0.85, 0.85, 1.0);
-	}
+    float line = float(int(gl_FragCoord.y));
+    if (int(mod(line, 2.0)) == 0) {
+        gl_FragColor = mix(rgb, vec4(1.0, 1.0, 1.0, 1.0), 0.05);
+    }
+    else {
+        gl_FragColor = rgb * vec4(0.85, 0.85, 0.85, 1.0);
+    }
 }
 
 ]]></fragment>
